complaint_ai/requirements.txt
# Complaint AI — Dependencies
# Install with: pip install -r requirements.txt

# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi==0.115.12
uvicorn[standard]==0.34.0

# ── Agent Orchestration ───────────────────────────────────────────────────────
langgraph==0.3.5

# ── LLM ──────────────────────────────────────────────────────────────────────
google-generativeai==0.8.4

# ── RAG / Vector Search ───────────────────────────────────────────────────────
faiss-cpu==1.9.0
sentence-transformers==3.4.1

# ── Database ──────────────────────────────────────────────────────────────────
pymongo==4.11.4

# ── Data Processing ───────────────────────────────────────────────────────────
pandas==2.2.3
numpy==1.26.4

# ── Config & Environment ──────────────────────────────────────────────────────
python-dotenv==1.0.1
